frontend/api/requirements.txt
reportlab==4.1.0
requests==2.31.0
google-generativeai==0.4.1
python-dotenv==1.0.1
python-docx==1.1.0
pdf2docx==0.5.8
